requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
supabase
docx
PyPDF2
langchain
langchain-community
langchain-text-splitters
langchain-openai
pydantic
pydantic-settings
dashscope
tencentcloud-sdk-python
psutil
redis
